+++ packages/emoji/CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## To be released
 
+- change EmojiSuggestions to popper.js with option `popperOptions` and deprecate `positionSuggestions` [TBD](https://github.com/draft-js-plugins/draft-js-plugins/issues/TBD)
+
 ## 4.6.6
 - added basic keyboard navigation to fix [#3233](https://github.com/draft-js-plugins/draft-js-plugins/issues/3233)